fix: restore p3 regression validation
     ## Summary
     - restore regression compatibility after P3 approval policy changes
     - mark read-only fake/test tools explicitly in older tests
     - update hook-only tests to bypass invocation approval policy
     - align spawn external-side-effect baseline with approval defer behavior
     - fix pyright narrowing for nested metadata and JSON-like traces

     ## Verification
     - pyright --level error: 0 errors
     - pyright --project pyrightconfig.tests.json --level error: 0 errors
     - pytest -q -p no:cacheprovider: 2003 passed, 3 warnings
     - compileall: passed
     - git diff --check: passed

+++ agent/task_plan/execution_runtime.py
@@ -134,7 +134,7 @@ class TaskExecutionLeaseGuard:
         *,
         session_key: str,
         attempt_id: str,
-        lease_seconds: int,
+        lease_seconds: float,
         clock: Callable[[], datetime],
     ) -> None:
         self._service = service
